relocate EventIndex
